Moving Structopts into separate module
