+++ src/main.rs
@@ -1,25 +1,24 @@
 #![allow(clippy::non_ascii_literal)]
 
 use crate::actions::new;
+use crate::cli_options::{ExternalCommand, ExternalSubcommands, Opt, SubCommand};
 use crate::dep_types::{Constraint, Lock, Package, Req, Version};
+use crate::pyproject::Config;
 use crate::util::abort;
 use crate::util::deps::sync;
 
 use regex::Regex;
-use serde::Deserialize;
 use std::{
-    collections::HashMap,
     env, fs,
     path::{Path, PathBuf},
-    str::FromStr,
     sync::{Arc, RwLock},
 };
 
-use structopt::StructOpt;
 use termcolor::{Color, ColorChoice};
 
 mod actions;
 mod build;
+mod cli_options;
 mod commands;
 mod dep_parser;
 mod dep_resolution;
@@ -27,6 +26,7 @@ mod dep_types;
 mod files;
 mod install;
 mod py_versions;
+mod pyproject;
 mod script;
 mod util;
 
@@ -40,190 +40,6 @@ mod util;
 // clear download git source as an option. In general, git install is a mess
 
 type PackToInstall = ((String, Version), Option<(u32, String)>); // ((Name, Version), (parent id, rename name))
-
-#[derive(StructOpt, Debug)]
-#[structopt(name = "pyflow", about = "Python packaging and publishing")]
-struct Opt {
-    #[structopt(subcommand)]
-    subcmds: SubCommand,
-
-    /// Force a color option: auto (default), always, ansi, never
-    #[structopt(short, long)]
-    color: Option<String>,
-}
-
-#[derive(StructOpt, Debug)]
-enum SubCommand {
-    /// Create a project folder with the basics
-    #[structopt(name = "new")]
-    New {
-        #[structopt(name = "name")]
-        name: String, // holds the project name.
-    },
-
-    /** Install packages from `pyproject.toml`, `pyflow.lock`, or specified ones. Example:
-
-    `pyflow install`: sync your installation with `pyproject.toml`, or `pyflow.lock` if it exists.
-    `pyflow install numpy scipy`: install `numpy` and `scipy`.*/
-    #[structopt(name = "install")]
-    Install {
-        #[structopt(name = "packages")]
-        packages: Vec<String>,
-        /// Save package to your dev-dependencies section
-        #[structopt(short, long)]
-        dev: bool,
-    },
-    /// Uninstall all packages, or ones specified
-    #[structopt(name = "uninstall")]
-    Uninstall {
-        #[structopt(name = "packages")]
-        packages: Vec<String>,
-    },
-    /// Display all installed packages and console scripts
-    #[structopt(name = "list")]
-    List,
-    /// Build the package - source and wheel
-    #[structopt(name = "package")]
-    Package {
-        #[structopt(name = "extras")]
-        extras: Vec<String>,
-    },
-    /// Publish to `pypi`
-    #[structopt(name = "publish")]
-    Publish,
-    /// Create a `pyproject.toml` from requirements.txt, pipfile etc, setup.py etc
-    #[structopt(name = "init")]
-    Init,
-    /// Remove the environment, and uninstall all packages
-    #[structopt(name = "reset")]
-    Reset,
-    /// Remove cached packages, Python installs, or script-environments. Eg to free up hard drive space.
-    #[structopt(name = "clear")]
-    Clear,
-    /// Run a CLI script like `ipython` or `black`. Note that you can simply run `pyflow black`
-    /// as a shortcut.
-    // Dummy option with space at the end for documentation
-    #[structopt(name = "run ")] // We don't need to invoke this directly, but the option exists
-    Run,
-
-    /// Run the project python or script with the project python environment.
-    /// As a shortcut you can simply specify a script name ending in `.py`
-    // Dummy option with space at the end for documentation
-    #[structopt(name = "python ")]
-    Python,
-
-    /// Run a standalone script not associated with a project
-    // Dummy option with space at the end for documentation
-    #[structopt(name = "script ")]
-    Script,
-    //    /// Run a package globally; used for CLI tools like `ipython` and `black`. Doesn't
-    //    /// interfere Python installations. Must have been installed with `pyflow install -g black` etc
-    //    #[structopt(name = "global")]
-    //    Global {
-    //        #[structopt(name = "name")]
-    //        name: String,
-    //    },
-    /// Change the Python version for this project. eg `pyflow switch 3.8`. Equivalent to setting
-    /// `py_version` in `pyproject.toml`.
-    #[structopt(name = "switch")]
-    Switch {
-        #[structopt(name = "version")]
-        version: String,
-    },
-    // Documentation for supported external subcommands can be documented by
-    // adding a `dummy` subcommand with the name having a trailing space.
-    // #[structopt(name = "external ")]
-    #[structopt(external_subcommand, name = "external")]
-    External(Vec<String>),
-}
-
-#[derive(Clone, Debug)]
-enum ExternalSubcommands {
-    Run,
-    Script,
-    Python,
-    ImpliedRun(String),
-    ImpliedPython(String),
-}
-
-impl ToString for ExternalSubcommands {
-    fn to_string(&self) -> String {
-        match self {
-            Self::Run => "run".into(),
-            Self::Script => "script".into(),
-            Self::Python => "python".into(),
-            Self::ImpliedRun(x) => x.into(),
-            Self::ImpliedPython(x) => x.into(),
-        }
-    }
-}
-
-impl FromStr for ExternalSubcommands {
-    type Err = anyhow::Error;
-    fn from_str(s: &str) -> anyhow::Result<Self> {
-        let result = match s {
-            "run" => Self::Run,
-            "script" => Self::Script,
-            "python" => Self::Python,
-            x if x.ends_with(".py") => Self::ImpliedPython(x.to_string()),
-            x => Self::ImpliedRun(x.to_string()),
-        };
-        Ok(result)
-    }
-}
-
-#[derive(Clone, Debug)]
-struct ExternalCommand {
-    cmd: ExternalSubcommands,
-    args: Vec<String>,
-}
-
-impl ExternalCommand {
-    fn from_opt(args: Vec<String>) -> Self {
-        let cmd = ExternalSubcommands::from_str(&args[0]).unwrap();
-        let cmd_args = match cmd {
-            ExternalSubcommands::Run
-            | ExternalSubcommands::Script
-            | ExternalSubcommands::Python => &args[1..],
-            ExternalSubcommands::ImpliedRun(_) | ExternalSubcommands::ImpliedPython(_) => &args,
-        };
-        let cmd = match cmd {
-            ExternalSubcommands::ImpliedRun(_) => ExternalSubcommands::Run,
-            ExternalSubcommands::ImpliedPython(_) => ExternalSubcommands::Python,
-            x => x,
-        };
-        Self {
-            cmd,
-            args: cmd_args.to_vec(),
-        }
-    }
-}
-/// A config, parsed from pyproject.toml
-#[derive(Clone, Debug, Default, Deserialize)]
-// todo: Auto-desr some of these
-pub struct Config {
-    name: Option<String>,
-    py_version: Option<Version>,
-    reqs: Vec<Req>,
-    dev_reqs: Vec<Req>,
-    version: Option<Version>,
-    authors: Vec<String>,
-    license: Option<String>,
-    extras: HashMap<String, String>,
-    description: Option<String>,
-    classifiers: Vec<String>, // https://pypi.org/classifiers/
-    keywords: Vec<String>,
-    homepage: Option<String>,
-    repository: Option<String>,
-    repo_url: Option<String>,
-    package_url: Option<String>,
-    readme: Option<String>,
-    build: Option<String>, // A python file used to build non-python extensions
-    //    entry_points: HashMap<String, Vec<String>>, // todo option?
-    scripts: HashMap<String, String>, //todo: put under [tool.pyflow.scripts] ?
-    //    console_scripts: Vec<String>, // We don't parse these; pass them to `setup.py` as-entered.
-    python_requires: Option<String>,
-}
 
 /// Reduce repetition between reqs and dev reqs when populating reqs of path reqs.
 fn pop_reqs_helper(reqs: &[Req], dev: bool) -> Vec<Req> {
@@ -271,375 +87,6 @@ fn pop_reqs_helper(reqs: &[Req], dev: bool) -> Vec<Req> {
         }
     }
     result
-}
-
-impl Config {
-    /// Helper fn to prevent repetition
-    fn parse_deps(deps: HashMap<String, files::DepComponentWrapper>) -> Vec<Req> {
-        let mut result = Vec::new();
-        for (name, data) in deps {
-            let constraints;
-            let mut extras = None;
-            let mut git = None;
-            let mut path = None;
-            let mut python_version = None;
-            match data {
-                files::DepComponentWrapper::A(constrs) => {
-                    constraints = if let Ok(c) = Constraint::from_str_multiple(&constrs) {
-                        c
-                    } else {
-                        abort(&format!(
-                            "Problem parsing constraints in `pyproject.toml`: {}",
-                            &constrs
-                        ));
-                        unreachable!()
-                    };
-                }
-                files::DepComponentWrapper::B(subdata) => {
-                    constraints = match subdata.constrs {
-                        Some(constrs) => {
-                            if let Ok(c) = Constraint::from_str_multiple(&constrs) {
-                                c
-                            } else {
-                                abort(&format!(
-                                    "Problem parsing constraints in `pyproject.toml`: {}",
-                                    &constrs
-                                ));
-                                unreachable!()
-                            }
-                        }
-                        None => vec![],
-                    };
-
-                    if let Some(ex) = subdata.extras {
-                        extras = Some(ex);
-                    }
-                    if let Some(p) = subdata.path {
-                        path = Some(p);
-                    }
-                    if let Some(repo) = subdata.git {
-                        git = Some(repo);
-                    }
-                    if let Some(v) = subdata.python {
-                        let pv = Constraint::from_str(&v)
-                            .expect("Problem parsing python version in dependency");
-                        python_version = Some(vec![pv]);
-                    }
-                }
-            }
-
-            result.push(Req {
-                name,
-                constraints,
-                extra: None,
-                sys_platform: None,
-                python_version,
-                install_with_extras: extras,
-                path,
-                git,
-            });
-        }
-        result
-    }
-
-    // todo: DRY at the top from `from_file`.
-    fn from_pipfile(path: &Path) -> Option<Self> {
-        // todo: Lots of tweaks and QC could be done re what fields to parse, and how best to
-        // todo parse and store them.
-        let toml_str = match fs::read_to_string(path).ok() {
-            Some(d) => d,
-            None => return None,
-        };
-
-        let decoded: files::Pipfile = if let Ok(d) = toml::from_str(&toml_str) {
-            d
-        } else {
-            abort("Problem parsing `Pipfile`");
-            unreachable!()
-        };
-        let mut result = Self::default();
-
-        if let Some(pipfile_deps) = decoded.packages {
-            result.reqs = Self::parse_deps(pipfile_deps);
-        }
-        if let Some(pipfile_dev_deps) = decoded.dev_packages {
-            result.dev_reqs = Self::parse_deps(pipfile_dev_deps);
-        }
-
-        Some(result)
-    }
-
-    /// Pull config data from `pyproject.toml`. We use this to deserialize things like Versions
-    /// and requirements.
-    fn from_file(path: &Path) -> Option<Self> {
-        // todo: Lots of tweaks and QC could be done re what fields to parse, and how best to
-        // todo parse and store them.
-        let toml_str = match fs::read_to_string(path) {
-            Ok(d) => d,
-            Err(_) => return None,
-        };
-
-        let decoded: files::Pyproject = if let Ok(d) = toml::from_str(&toml_str) {
-            d
-        } else {
-            abort("Problem parsing `pyproject.toml`");
-            unreachable!()
-        };
-        let mut result = Self::default();
-
-        // Parse Poetry first, since we'll use pyflow if there's a conflict.
-        if let Some(po) = decoded.tool.poetry {
-            if let Some(v) = po.name {
-                result.name = Some(v);
-            }
-            if let Some(v) = po.authors {
-                result.authors = v;
-            }
-            if let Some(v) = po.license {
-                result.license = Some(v);
-            }
-
-            if let Some(v) = po.homepage {
-                result.homepage = Some(v);
-            }
-            if let Some(v) = po.description {
-                result.description = Some(v);
-            }
-            if let Some(v) = po.repository {
-                result.repository = Some(v);
-            }
-            if let Some(v) = po.readme {
-                result.readme = Some(v);
-            }
-            if let Some(v) = po.build {
-                result.build = Some(v);
-            }
-            // todo: Process entry pts, classifiers etc?
-            if let Some(v) = po.classifiers {
-                result.classifiers = v;
-            }
-            if let Some(v) = po.keywords {
-                result.keywords = v;
-            }
-
-            //                        if let Some(v) = po.source {
-            //                result.source = v;
-            //            }
-            //            if let Some(v) = po.scripts {
-            //                result.console_scripts = v;
-            //            }
-            if let Some(v) = po.extras {
-                result.extras = v;
-            }
-
-            if let Some(v) = po.version {
-                result.version = Some(
-                    Version::from_str(&v).expect("Problem parsing version in `pyproject.toml`"),
-                )
-            }
-
-            // todo: DRY (c+p) from pyflow dependency parsing, other than parsing python version here,
-            // todo which only poetry does.
-            // todo: Parse poetry dev deps
-            if let Some(deps) = po.dependencies {
-                for (name, data) in deps {
-                    let constraints;
-                    let mut extras = None;
-                    let mut python_version = None;
-                    match data {
-                        files::DepComponentWrapperPoetry::A(constrs) => {
-                            constraints = Constraint::from_str_multiple(&constrs)
-                                .expect("Problem parsing constraints in `pyproject.toml`.");
-                        }
-                        files::DepComponentWrapperPoetry::B(subdata) => {
-                            constraints = Constraint::from_str_multiple(&subdata.constrs)
-                                .expect("Problem parsing constraints in `pyproject.toml`.");
-                            if let Some(ex) = subdata.extras {
-                                extras = Some(ex);
-                            }
-                            if let Some(v) = subdata.python {
-                                let pv = Constraint::from_str(&v)
-                                    .expect("Problem parsing python version in dependency");
-                                python_version = Some(vec![pv]);
-                            }
-                            // todo repository etc
-                        }
-                    }
-                    if &name.to_lowercase() == "python" {
-                        if let Some(constr) = constraints.get(0) {
-                            result.py_version = Some(constr.version.clone())
-                        }
-                    } else {
-                        result.reqs.push(Req {
-                            name,
-                            constraints,
-                            extra: None,
-                            sys_platform: None,
-                            python_version,
-                            install_with_extras: extras,
-                            path: None,
-                            git: None,
-                        });
-                    }
-                }
-            }
-        }
-
-        if let Some(pf) = decoded.tool.pyflow {
-            if let Some(v) = pf.name {
-                result.name = Some(v);
-            }
-
-            if let Some(v) = pf.authors {
-                result.authors = if v.is_empty() {
-                    util::get_git_author()
-                } else {
-                    v
-                };
-            }
-            if let Some(v) = pf.license {
-                result.license = Some(v);
-            }
-            if let Some(v) = pf.homepage {
-                result.homepage = Some(v);
-            }
-            if let Some(v) = pf.description {
-                result.description = Some(v);
-            }
-            if let Some(v) = pf.repository {
-                result.repository = Some(v);
-            }
-
-            // todo: Process entry pts, classifiers etc?
-            if let Some(v) = pf.classifiers {
-                result.classifiers = v;
-            }
-            if let Some(v) = pf.keywords {
-                result.keywords = v;
-            }
-            if let Some(v) = pf.readme {
-                result.readme = Some(v);
-            }
-            if let Some(v) = pf.build {
-                result.build = Some(v);
-            }
-            //            if let Some(v) = pf.entry_points {
-            //                result.entry_points = v;
-            //            } // todo
-            if let Some(v) = pf.scripts {
-                result.scripts = v;
-            }
-
-            if let Some(v) = pf.python_requires {
-                result.python_requires = Some(v);
-            }
-
-            if let Some(v) = pf.package_url {
-                result.package_url = Some(v);
-            }
-
-            if let Some(v) = pf.version {
-                result.version = Some(
-                    Version::from_str(&v).expect("Problem parsing version in `pyproject.toml`"),
-                )
-            }
-
-            if let Some(v) = pf.py_version {
-                result.py_version = Some(
-                    Version::from_str(&v)
-                        .expect("Problem parsing python version in `pyproject.toml`"),
-                );
-            }
-
-            if let Some(deps) = pf.dependencies {
-                result.reqs = Self::parse_deps(deps);
-            }
-            if let Some(deps) = pf.dev_dependencies {
-                result.dev_reqs = Self::parse_deps(deps);
-            }
-        }
-
-        Some(result)
-    }
-
-    /// For reqs of `path` type, add their sub-reqs by parsing `setup.py` or `pyproject.toml`.
-    fn populate_path_subreqs(&mut self) {
-        self.reqs.append(&mut pop_reqs_helper(&self.reqs, false));
-        self.dev_reqs
-            .append(&mut pop_reqs_helper(&self.dev_reqs, true));
-    }
-
-    /// Create a new `pyproject.toml` file.
-    fn write_file(&self, path: &Path) {
-        let file = path;
-        if file.exists() {
-            abort("`pyproject.toml` already exists")
-        }
-
-        let mut result = String::new();
-
-        result.push_str("\n[tool.pyflow]\n");
-        if let Some(name) = &self.name {
-            result.push_str(&("name = \"".to_owned() + name + "\"\n"));
-        } else {
-            // Give name, and a few other fields default values.
-            result.push_str(&("name = \"\"".to_owned() + "\n"));
-        }
-        if let Some(py_v) = &self.py_version {
-            result.push_str(&("py_version = \"".to_owned() + &py_v.to_string_no_patch() + "\"\n"));
-        } else {
-            result.push_str(&("py_version = \"3.8\"".to_owned() + "\n"));
-        }
-        if let Some(vers) = self.version.clone() {
-            result.push_str(&(format!("version = \"{}\"", vers.to_string() + "\n")));
-        } else {
-            result.push_str("version = \"0.1.0\"");
-            result.push('\n');
-        }
-        if !self.authors.is_empty() {
-            result.push_str("authors = [\"");
-            for (i, author) in self.authors.iter().enumerate() {
-                if i != 0 {
-                    result.push_str(", ");
-                }
-                result.push_str(author);
-            }
-            result.push_str("\"]\n");
-        }
-
-        if let Some(v) = &self.description {
-            result.push_str(&(format!("description = \"{}\"", v) + "\n"));
-        }
-        if let Some(v) = &self.homepage {
-            result.push_str(&(format!("homepage = \"{}\"", v) + "\n"));
-        }
-
-        // todo: More fields
-
-        result.push('\n');
-        result.push_str("[tool.pyflow.scripts]\n");
-        for (name, mod_fn) in &self.scripts {
-            result.push_str(&(format!("{} = \"{}\"", name, mod_fn) + "\n"));
-        }
-
-        result.push('\n');
-        result.push_str("[tool.pyflow.dependencies]\n");
-        for dep in &self.reqs {
-            result.push_str(&(dep.to_cfg_string() + "\n"));
-        }
-
-        result.push('\n');
-        result.push_str("[tool.pyflow.dev-dependencies]\n");
-        for dep in &self.dev_reqs {
-            result.push_str(&(dep.to_cfg_string() + "\n"));
-        }
-
-        result.push('\n'); // trailing newline
-
-        if fs::write(file, result).is_err() {
-            abort("Problem writing `pyproject.toml`")
-        }
-    }
 }
 
 /// Cli Config to hold command line options
@@ -711,7 +158,6 @@ fn run_cli_tool(
         a.clone()
     } else {
         abort("`run` must be followed by the script to run, eg `pyflow run black`");
-        unreachable!()
     };
 
     // If the script we're calling is specified in `pyproject.toml`, ensure it exists.
@@ -743,7 +189,6 @@ fn run_cli_tool(
             }
         } else {
             abort(&format!("Problem parsing the following script: {:#?}. Must be in the format module:function_name", s));
-            unreachable!()
         }
         return;
     }
@@ -837,10 +282,10 @@ const LOCK_FILENAME: &str = "pyflow.lock";
 /// We process input commands in a deliberate order, to ensure the required, and only the required
 /// setup steps are accomplished before each.
 fn main() {
-    let (pyflow_path, dep_cache_path, script_env_path, git_path) = util::fs::get_paths();
+    let (pyflow_path, dep_cache_path, script_env_path, git_path) = util::paths::get_paths();
     let os = util::get_os();
 
-    let opt = Opt::from_args();
+    let opt = <Opt as structopt::StructOpt>::from_args();
     #[cfg(debug_assertions)]
     eprintln!("opts {:?}", opt);
 
@@ -861,49 +306,53 @@ fn main() {
         None
     };
 
-    // Run this before parsing the config.
-    if let Some(x) = extcmd.clone() {
-        if let ExternalSubcommands::Script = x.cmd {
-            script::run_script(&script_env_path, &dep_cache_path, os, &x.args, &pyflow_path);
+    match &subcmd {
+        SubCommand::External(ref x) => match ExternalCommand::from_opt(x.to_owned()) {
+            ExternalCommand { cmd, args } => match cmd {
+                ExternalSubcommands::Script => {
+                    script::run_script(&script_env_path, &dep_cache_path, os, &args, &pyflow_path);
+                }
+                // TODO: Move branches to omitted match
+                _ => (),
+            },
+        },
+        SubCommand::New { name } => {
+            if new(&name).is_err() {
+                abort(actions::NEW_ERROR_MESSAGE);
+            }
+
+            util::print_color(
+                &format!("Created a new Python project named {}", name),
+                Color::Green,
+            );
             return;
         }
-    }
+        SubCommand::Init => {
+            let cfg_path = PathBuf::from(CFG_FILENAME);
+            if cfg_path.exists() {
+                abort("pyproject.toml already exists - not overwriting.")
+            }
 
-    if let SubCommand::New { name } = subcmd {
-        if new(&name).is_err() {
-            abort(actions::NEW_ERROR_MESSAGE);
+            let mut cfg = match PathBuf::from("Pipfile").exists() {
+                true => Config::from_pipfile(&PathBuf::from("Pipfile")).unwrap_or_default(),
+                false => Config::default(),
+            };
+
+            cfg.py_version = Some(util::prompts::py_vers());
+
+            files::parse_req_dot_text(&mut cfg, &PathBuf::from("requirements.txt"));
+
+            cfg.write_file(&cfg_path);
+            util::print_color("Created `pyproject.toml`", Color::Green);
+            // Don't return here; let the normal logic create the venv now.
         }
-        util::print_color(
-            &format!("Created a new Python project named {}", name),
-            Color::Green,
-        );
-        return;
-    }
-
-    if let SubCommand::Init {} = subcmd {
-        let cfg_path = PathBuf::from(CFG_FILENAME);
-        if cfg_path.exists() {
-            abort("pyproject.toml already exists - not overwriting.")
-        }
-
-        let mut cfg = match PathBuf::from("Pipfile").exists() {
-            true => Config::from_pipfile(&PathBuf::from("Pipfile")).unwrap_or_default(),
-            false => Config::default(),
-        };
-
-        cfg.py_version = Some(util::prompts::py_vers());
-
-        files::parse_req_dot_text(&mut cfg, &PathBuf::from("requirements.txt"));
-
-        cfg.write_file(&cfg_path);
-        util::print_color("Created `pyproject.toml`", Color::Green);
-        // Don't return here; let the normal logic create the venv now.
+        // TODO: Move branches to omitted match
+        _ => {}
     }
 
     // We need access to the config from here on; throw an error if we can't find it.
     let mut cfg_path = PathBuf::from(CFG_FILENAME);
     if !&cfg_path.exists() {
-        //        if let SubCommand::Python { args: _ } = subcmd {
         // Try looking recursively in parent directories for a config file.
         let recursion_limit = 8; // How my levels to look up
         let mut current_level = env::current_dir().expect("Can't access current directory");
